+++ src/plugins/blocksmith/components/document-block-modal/index.jsx
@@ -159,8 +159,8 @@ class DocumentBlockModal extends React.Component {
     this.setState(changedTime(time));
   }
 
-  handleCloseUploader(assets) {
-    if (assets) {
+  handleCloseUploader(document) {
+    if (document) {
       this.handleSearchDocumentAssets();
     }
   }
@@ -169,11 +169,15 @@ class DocumentBlockModal extends React.Component {
     this.setState({ selectedImageNode: null }, this.refocusModal);
   }
 
-  handleToggleUploader() {
+  handleToggleUploader(document) {
     this.setState(({ isUploaderOpen }) => ({ isUploaderOpen: !isUploaderOpen }), () => {
       const { isUploaderOpen } = this.state;
       if (!isUploaderOpen) {
         this.refocusModal();
+        if (document) {
+          this.handleSearchDocumentAssets();
+          this.handleSelectDocument(document);
+        }
       }
     });
   }